Jehyun

+++ Back/readforce/src/main/java/com/readforce/dto/PointDto.java
@@ -25,17 +25,17 @@ public class PointDto {
 	@AllArgsConstructor
 	public static class PointRanking{
 
-		private Long total = 0L;
+		private Double total = 0.0;
 
-		private Long korean_news = 0L;
+		private Double korean_news = 0.0;
 
-		private Long english_news = 0L;
+		private Double english_news = 0.0;
 
-		private Long japanese_news = 0L;
+		private Double japanese_news = 0.0;
 
-		private Long novel = 0L;
+		private Double novel = 0.0;
 
-		private Long fairytale = 0L;
+		private Double fairytale = 0.0;
 		
 		private String nickname;
 		
@@ -64,22 +64,22 @@ public class PointDto {
 		private String email;	
 		
 		@Min(value = 0, message = MessageCode.TOTAL_POINT_MUST_BE_POSITIVE_NUMBER)
-		private Long total = 0L;
+		private Double total = 0.0;
 
 		@Min(value = 0, message = MessageCode.KOREAN_NEWS_POINT_MUST_BE_POSITIVE_NUMBER)
-		private Long korean_news = 0L;
+		private Double korean_news = 0.0;
 
 		@Min(value = 0, message = MessageCode.ENGILISH_NEWS_POINT_MUST_BE_POSITIVE_NUMBER)
-		private Long english_news = 0L;
+		private Double english_news = 0.0;
 
 		@Min(value = 0, message = MessageCode.JAPANESE_NEWS_POINT_MUST_BE_POSITIVE_NUMBER)
-		private Long japanese_news = 0L;
+		private Double japanese_news = 0.0;
 
 		@Min(value = 0, message = MessageCode.NOVEL_POINT_MUST_BE_POSITIVE_NUMBER)
-		private Long novel = 0L;
+		private Double novel = 0.0;
 
 		@Min(value = 0, message = MessageCode.FAIRYTALE_POINT_MUST_BE_POSITIVE_NUMBER)
-		private Long fairytale = 0L;
+		private Double fairytale = 0.0;
 		
 	}
 	
@@ -91,17 +91,17 @@ public class PointDto {
 
 		private String email;
 
-		private Long total = 0L;
+		private Double total = 0.0;
 
-		private Long korean_news = 0L;
+		private Double korean_news = 0.0;
 
-		private Long english_news = 0L;
+		private Double english_news = 0.0;
 
-		private Long japanese_news = 0L;
+		private Double japanese_news = 0.0;
 
-		private Long novel = 0L;
+		private Double novel = 0.0;
 
-		private Long fairytale = 0L;
+		private Double fairytale = 0.0;
 
 		private LocalDateTime created_date;
